requirements.txt
streamlit-mic-recorder
InstructorEmbedding
ctransformers
transformers
langchain
streamlit
pypdfium2
librosa
pyyaml
chromadb
pydub